Add small test to make sure the terminal bot doesn't cause more message saves than is needed
